Better UPS Calculations
The update sequence for the environment needs to load the function on the GPU, which takes up time that will never been required in future updates. Due to this, the first update timer was skewing the results of the UPS calculations, so I simply removed the first element from the calculations.