Full implementation of finding one food patch
Removed ALL Neural Network processes as it was not feasible.

Food sources can be placed, as well as wall objects, and the agents will start their search for the food source. Once it's found, most agents (other than the ones who are lost) will quickly follow onto that pheromone path to the food. The lost ones will eventually find it too, but takes much longer.

When a more optimal path appears due to the wall obstruction opening up, the ants will eventually find that new path, assuming it is close by to their current path. This is not a necessary test point though, as it is not planned for this to occur in the final product