Food-to-food connections
Agents can now create trails between two different food sources, as well as multiple nest-to-food connections.

This is achieve through a trail popularity selection, where an agent chooses an adjacent cell with the greatest concentration of the pheromone it's following, while also having the lowest concentration of the opposing pheromone.